ehrsimulator/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
langchain
httpx
python-dotenv
sse-starlette
psycopg2-binary
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp-proto-http 